LayoutTests/platform/ios-sim-deprecated/http/tests/security/no-indexeddb-from-sandbox-expected.txt

ALERT: FAIL: db.open() threw a TypeError
